+++ R/ee_check.R
@@ -9,12 +9,14 @@
 #' @importFrom crayon yellow
 #' @importFrom cli cat_line
 #' @examples
+#' \dontrun{
 #' library(rgee)
 #' ee_reattach() # reattach ee as a reserved word
 #' ee_check_python()
 #' ee_check_rgee_python_packages()
 #' ee_check_credentials()
 #' ee_check() # put it all together
+#' }
 #' @export
 ee_check <- function() {
   ee_check_python()
@@ -87,7 +89,7 @@ ee_check_rgee_python_packages <- function(quiet = FALSE) {
         ". To avoid possible issues, we ",
         "highly recommend install the version used by rgee ",
         ee_version(),
-        ", you might use:\n >>> ee_install_python_packages() \n",
+        ", you might use:\n >>> ee_earthengine_upgrade() \n",
         " >>> pip install earthengine-api==",ee_version(),
         "\n >>> conda install earthengine-api==",ee_version(),
         "\nIf the installation is successful, restart to see changes."