+++ tests/testthat/test-map.R
@@ -1,7 +1,15 @@
 context("rgee: sf_as_ee test")
 
+if (isFALSE(exists('ee'))) {
+  ee_reattach()
+  ee_Initialize(
+    email = 'data.colec.fbf@gmail.com',
+    drive = TRUE,
+    gcs = TRUE
+  )
+}
+
 # data --------------------------------------------------------------------
-ee_Initialize(email = 'data.colec.fbf@gmail.com', drive = TRUE, gcs = TRUE)
 geom <- ee$Geometry$Point(list(-73.53522, -15.75453))
 eeobject_fc <- ee$FeatureCollection("users/csaybar/DLdemos/train_set")
 image <- ee$Image("LANDSAT/LC08/C01/T1/LC08_044034_20140318")